refactor: make layout folder
